refactor: split convergence utils
