+++ tests/test_cl_cd_stats.py
@@ -3,7 +3,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
 import numpy as np
-from glacium.utils.convergence import cl_cd_stats
+from glacium.utils.convergence.stats import cl_cd_stats
 
 
 def test_cl_cd_stats(tmp_path):
